changes
After user login . they there will be practice and test. I'm going to create a clickable screen for tobe and past tense . user should be able to click the tobe it will get them a question about to be and if they choose past tense it will give them that question .  if  the user click the practice section i should get them into that screen i create .I also fix the score and progress.